fixed user components and moved each component seperately.
