+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import User from './users/users';
+import UserHeader from './users/userHeader';
 import './scss/styles.scss';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <User id="XXX" />
+        <UserHeader id="XXX" />
       </div>
     );
   }
