+++ src/scss/typography.scss
@@ -1,0 +1,3 @@
+a, a:link {
+    text-decoration: none;
+}
