+++ src/scss/user.scss
@@ -1,8 +1,29 @@
 .UserHead {
   display: flex;
+  padding: 5px;
 
     .Avatar {
-        width:50px;
-        height: 50px;
+        width:40px;
+        height: 40px;
+        border-radius: 50%;
+        overflow: hidden;
+        float: left;
     }
+}
+
+.UserInfo {
+    color: #365899;
+    cursor: pointer;
+    text-decoration: none;
+    font-weight: bold;
+    text-decoration: none;
+    display:inline;
+    padding: 5px 10px 0px;
+}
+
+.timeLabel {
+    font-size: 12px;
+    color: #616770;
+    display: block;
+    padding: 0px 49px;
 }
